Split libtorch.so back into libtorch_{cpu,cuda,hip}
The new structure is that libtorch_cpu contains the bulk of our
code, and libtorch depends on libtorch_cpu and libtorch_cuda.

There's probably going to be quite a bit of one-off fixing;
I'll fill in all the details when I get it to work.

Fixes #27215 (as our libraries are smaller), and executes on
part of the plan in #29235.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 298075c877717f21e99fd4857f4b26a45ed2ced3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29731